Patch BAR files

+++ OpenKh.Tests/Patcher/PatcherTests.cs
@@ -1,7 +1,10 @@
+using OpenKh.Common;
+using OpenKh.Kh2;
 using OpenKh.Patcher;
 using System;
 using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 using Xunit;
 using Xunit.Sdk;
 
@@ -15,6 +18,7 @@ namespace OpenKh.Tests.Patcher
 
         public PatcherTests()
         {
+            Dispose();
             Directory.CreateDirectory(AssetsInputDir);
             Directory.CreateDirectory(ModInputDir);
             Directory.CreateDirectory(ModOutputDir);
@@ -22,9 +26,12 @@ namespace OpenKh.Tests.Patcher
 
         public void Dispose()
         {
-            Directory.Delete(AssetsInputDir, true);
-            Directory.Delete(ModInputDir, true);
-            Directory.Delete(ModOutputDir, true);
+            if (Directory.Exists(AssetsInputDir))
+                Directory.Delete(AssetsInputDir, true);
+            if (Directory.Exists(ModInputDir))
+                Directory.Delete(ModInputDir, true);
+            if (Directory.Exists(ModOutputDir))
+                Directory.Delete(ModOutputDir, true);
         }
 
         [Fact]
@@ -55,11 +62,162 @@ namespace OpenKh.Tests.Patcher
             AssertFileExists(ModOutputDir, patch.Assets.Kh2.Binaries[0].Name);
         }
 
+        [Fact]
+        public void Kh2CreateBinArcIfSourceDoesntExistsTest()
+        {
+            var patcher = new PatcherProcessor();
+            var patch = new Metadata
+            {
+                Assets = new AssetContainer
+                {
+                    Kh2 = new AssetKh2
+                    {
+                        BinaryArchives = new List<AssetBinArc>
+                        {
+                            new AssetBinArc
+                            {
+                                Name = "somedir/somefile.bar",
+                                Entries = new List<AssetFile>
+                                {
+                                    new AssetFile
+                                    {
+                                        Name = "abcd",
+                                        Format = "list",
+                                        Method = "copy",
+                                        Source = new AssetSource
+                                        {
+                                            Name = "somedir/somefile/abcd.bin"
+                                        }
+                                    }
+                                }
+                            }
+                        }
+                    }
+                }
+            };
+
+            CreateFile(ModInputDir, "somedir/somefile/abcd.bin").Using(x =>
+            {
+                x.WriteByte(0);
+                x.WriteByte(1);
+                x.WriteByte(2);
+                x.WriteByte(3);
+            });
+
+            patcher.Patch(AssetsInputDir, ModOutputDir, patch, ModInputDir);
+
+            AssertFileExists(ModOutputDir, patch.Assets.Kh2.BinaryArchives[0].Name);
+            AssertBarFile("abcd", entry =>
+            {
+                Assert.Equal(Bar.EntryType.List, entry.Type);
+                Assert.Equal(4, entry.Stream.Length);
+                Assert.Equal(0, entry.Stream.ReadByte());
+                Assert.Equal(1, entry.Stream.ReadByte());
+                Assert.Equal(2, entry.Stream.ReadByte());
+                Assert.Equal(3, entry.Stream.ReadByte());
+            }, ModOutputDir, patch.Assets.Kh2.BinaryArchives[0].Name);
+        }
+
+        [Fact]
+        public void Kh2MergeWithOriginalBinArcTest()
+        {
+            var patcher = new PatcherProcessor();
+            var patch = new Metadata
+            {
+                Assets = new AssetContainer
+                {
+                    Kh2 = new AssetKh2
+                    {
+                        BinaryArchives = new List<AssetBinArc>
+                        {
+                            new AssetBinArc
+                            {
+                                Name = "somedir/somefile.bar",
+                                Entries = new List<AssetFile>
+                                {
+                                    new AssetFile
+                                    {
+                                        Name = "abcd",
+                                        Format = "list",
+                                        Method = "copy",
+                                        Source = new AssetSource
+                                        {
+                                            Name = "somedir/somefile/abcd.bin"
+                                        }
+                                    }
+                                }
+                            }
+                        }
+                    }
+                }
+            };
+
+            CreateFile(ModInputDir, "somedir/somefile/abcd.bin").Using(x =>
+            {
+                x.WriteByte(0);
+                x.WriteByte(1);
+                x.WriteByte(2);
+                x.WriteByte(3);
+            });
+
+            CreateFile(AssetsInputDir, "somedir/somefile.bar").Using(x =>
+            {
+                Bar.Write(x, new Bar
+                {
+                    new Bar.Entry
+                    {
+                        Name = "nice",
+                        Type = Bar.EntryType.Model,
+                        Stream = new MemoryStream(new byte[] { 4, 5, 6, 7 })
+                    }
+                });
+            });
+
+            patcher.Patch(AssetsInputDir, ModOutputDir, patch, ModInputDir);
+
+            AssertFileExists(ModOutputDir, patch.Assets.Kh2.BinaryArchives[0].Name);
+            AssertBarFile("abcd", entry =>
+            {
+                Assert.Equal(Bar.EntryType.List, entry.Type);
+                Assert.Equal(4, entry.Stream.Length);
+                Assert.Equal(0, entry.Stream.ReadByte());
+                Assert.Equal(1, entry.Stream.ReadByte());
+                Assert.Equal(2, entry.Stream.ReadByte());
+                Assert.Equal(3, entry.Stream.ReadByte());
+            }, ModOutputDir, patch.Assets.Kh2.BinaryArchives[0].Name);
+            AssertBarFile("nice", entry =>
+            {
+                Assert.Equal(Bar.EntryType.Model, entry.Type);
+                Assert.Equal(4, entry.Stream.Length);
+                Assert.Equal(4, entry.Stream.ReadByte());
+                Assert.Equal(5, entry.Stream.ReadByte());
+                Assert.Equal(6, entry.Stream.ReadByte());
+                Assert.Equal(7, entry.Stream.ReadByte());
+            }, ModOutputDir, patch.Assets.Kh2.BinaryArchives[0].Name);
+        }
+
         private static void AssertFileExists(params string[] paths)
         {
             var filePath = Path.Join(paths);
             if (File.Exists(filePath) == false)
-                throw new TrueException($"Expected file to be found: {filePath}", false);
+                throw new TrueException($"File not found '{filePath}'", false);
+        }
+
+        private static void AssertBarFile(string name, Action<Bar.Entry> assertion, params string[] paths)
+        {
+            var filePath = Path.Join(paths);
+            var entries = File.OpenRead(filePath).Using(x =>
+            {
+                if (!Bar.IsValid(x))
+                    throw new TrueException($"Not a valid BinArc", false);
+                return Bar.Read(x);
+            });
+
+            var entry = entries.SingleOrDefault(x => x.Name == name);
+            if (entry == null)
+                throw new XunitException($"Entry '{name}' not found");
+
+            assertion(entry);
         }
 
         private static Stream CreateFile(params string[] paths)
